# backend/src/api/Users.ts
import * as express from "express";
import * as mongoose from "mongoose";
import * as logger from "winston";
import * as SerializationConverter from "../convertSrdToBooleanLists/SerializationConverter";
import {fromJSON} from "../nodes/JsonParser";
import {addApiToRequest} from "../spotify/Authorization";
import {Playlist as SpotifyPlaylist} from "../spotify/Playlist";
import {InitializedSpotifyApi} from "../spotify/SpotifyApi";

mongoose.connect(process.env.MONGOOSE_CONNECTION_STRING);

const Schema = mongoose.Schema;

const Playlist = new Schema({
    description: String,
    graph: String,
    name: String,
    uri: String,
});

const UserSchema = new Schema({
    authorization: {
        accessToken: String,
        expiresAt: Number,
        refreshToken: String,
    },
    id: String,
    playlists: [Playlist],
}, {strict: true});

export const User = mongoose.model("User", UserSchema);

export async function getOrCreateUser(id: string) {
    let user = await User.findOne({id});
    if (!user) {
        user = User.create({id});
    }

    return user;
}

const router: express.Router = express.Router();

interface IPlaylist {
    description: string;
    graph: string;
    name: string;
    uri: string;
}

router.get("/", async (req, res) => {
    const id: string = (req as any).kauth.grant.access_token.content.sub;
    const user = await getOrCreateUser(id);

    const connectedToSpotify: boolean = user.authorization && Date.now() < user.authorization.expiresAt;

    res.json({
        connectedToSpotify,
    });
});

router.get("/playlist", async (req, res) => {
    const id: string = (req as any).kauth.grant.access_token.content.sub;
    const user = await getOrCreateUser(id);

    const playlists: IPlaylist[] = [];
    for (const playlist of user.playlists) {
        if (playlist.name && playlist.graph) {
            playlists.push({
                description: playlist.description,
                graph: playlist.graph,
                name: playlist.name,
                uri: playlist.uri,
            });
        }
    }

    res.json({
        elements: playlists.length,
        elementsPerPage: playlists.length,
        page: 0,
        playlists,
    });
});

router.get("/playlist/:id", async (req, res) => {
    const id: string = (req as any).kauth.grant.access_token.content.sub;
    const user = await getOrCreateUser(id);

    const playlist = findPlaylist(user, req.params.id);
    if (playlist) {
        res.json({
            description: playlist.description,
            graph: JSON.parse(playlist.graph),
            name: playlist.name,
            uri: playlist.uri,
        });
    } else {
        res.sendStatus(404);
    }
});

function findPlaylist(user: any, name: string) {
    return user.playlists.find((playlistItem: any) => playlistItem.name === name);
}

export async function savePlaylist(userId: string, playlist: IPlaylist, uri?: string) {
    logger.info(`Searching user with id ${userId}`);
    const user = await getOrCreateUser(userId);
    const graph = typeof playlist.graph !== "string" ? JSON.stringify(playlist.graph) : playlist.graph;
    if (uri) {
        playlist.uri = uri;
    }

    const playlistEntity = findPlaylist(user, playlist.name);
    if (playlistEntity) {
        playlistEntity.description = playlist.description;
        playlistEntity.graph = graph;
        playlistEntity.name = playlist.name;
        playlistEntity.uri = playlist.uri;
    } else {
        user.playlists.push({
            description: playlist.description,
            graph,
            name: playlist.name,
            uri: playlist.uri,
        });
    }
    await user.save();
}

async function savePlaylistToSpotify(req: any, res: any) {
    try {
        await addApiToRequest(req);
        const api: InitializedSpotifyApi = new InitializedSpotifyApi((req as any).api);
        const serialized = SerializationConverter.convertSrdToBooleanList(req.body.graph);
        logger.info("Using parsed node", serialized);
        const node = await fromJSON(api, serialized);

        const tracksToAdd = await node.getTracks();
        logger.info(`Adding ${tracksToAdd.length} tracks.`);

        const me = await api.me();

        let playlist: SpotifyPlaylist;
        if (req.body.uri) {
            playlist = await SpotifyPlaylist.fromSpotifyUri(api, await me.id(), req.body.uri);
            if (playlist.name() !== req.body.name) {
                playlist = await me.createPlaylist(req.body.name);
            }
        } else if (req.body.name) {
            playlist = await me.createPlaylist(req.body.name);
        } else {
            throw new Error("Enter a name of a playlist");
        }

        await playlist.clear();
        await playlist.addTracks(tracksToAdd);

        res.json({message: "Successfully added songs.", playlistUri: playlist.id()});
    } catch (error) {
        logger.error(error.stack);
        res.sendStatus(500);
    }
}

router.post("/playlist", async (req, res) => {
    const id: string = (req as any).kauth.grant.access_token.content.sub;
    try {
        await savePlaylist(id, req.body);
        if (req.body.saveToSpotify) {
            await savePlaylistToSpotify(req, res);
        } else {
            res.json({});
        }
    } catch (error) {
        logger.error(error);
        res.sendStatus(500);
    }
});

router.delete("/playlist/:id", async (req, res) => {
    const id: string = (req as any).kauth.grant.access_token.content.sub;
    logger.info(`Searching user with id ${id}`);
    const user = await getOrCreateUser(id);

    const playlist = findPlaylist(user, req.params.id);
    if (playlist) {
        logger.info(`Found playlist ${req.params.id}, deleting it`);
        await playlist.remove();
        await user.save();
    }

    res.sendStatus(200);
});

export default router;